+++ cinemabe/Cinema/2-Business/Cinema.Business.DTO/Catalog/FoodAndDrinkDTOs.cs
@@ -6,7 +6,7 @@ public class FoodAndDrinkDTO
 {
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public string? ImageUrl { get; set; }
     public string? Description { get; set; }
     public bool IsAvailable { get; set; }
@@ -15,7 +15,7 @@ public class FoodAndDrinkDTO
 public class CreateFoodAndDrinkRequest
 {
     public string Name { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public string? ImageUrl { get; set; }
     public string? Description { get; set; }
     public bool IsAvailable { get; set; } = true;
@@ -25,7 +25,7 @@ public class UpdateFoodAndDrinkRequest : IHasId
 {
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal Price { get; set; }
+    public double Price { get; set; }
     public string? ImageUrl { get; set; }
     public string? Description { get; set; }
     public bool IsAvailable { get; set; } = true;
