+++ cinemabe/Cinema/2-Business/Cinema.Business.DTO/Catalog/HolidayDTOs.cs
@@ -7,14 +7,14 @@ public class HolidayDTO
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public DateOnly Date { get; set; }
-    public decimal PriceMultiplier { get; set; }
+    public double PriceMultiplier { get; set; }
 }
 
 public class CreateHolidayRequest
 {
     public string Name { get; set; } = string.Empty;
     public DateOnly Date { get; set; }
-    public decimal PriceMultiplier { get; set; } = 1.5m;
+    public double PriceMultiplier { get; set; } = 1.5;
 }
 
 public class UpdateHolidayRequest : IHasId
@@ -22,5 +22,5 @@ public class UpdateHolidayRequest : IHasId
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
     public DateOnly Date { get; set; }
-    public decimal PriceMultiplier { get; set; } = 1.5m;
+    public double PriceMultiplier { get; set; } = 1.5;
 }
