+++ cinemabe/Cinema/2-Business/Cinema.Business.DTO/Catalog/MemberShipDTOs.cs
@@ -8,7 +8,7 @@ public class MemberShipDTO
     public string Name { get; set; } = string.Empty;
     public int MinPoints { get; set; }
     public int MaxPoints { get; set; }
-    public decimal DiscountPercent { get; set; }
+    public double DiscountPercent { get; set; }
 }
 
 public class CreateMemberShipRequest
@@ -16,7 +16,7 @@ public class CreateMemberShipRequest
     public string Name { get; set; } = string.Empty;
     public int MinPoints { get; set; }
     public int MaxPoints { get; set; }
-    public decimal DiscountPercent { get; set; }
+    public double DiscountPercent { get; set; }
 }
 
 public class UpdateMemberShipRequest : IHasId
@@ -25,5 +25,5 @@ public class UpdateMemberShipRequest : IHasId
     public string Name { get; set; } = string.Empty;
     public int MinPoints { get; set; }
     public int MaxPoints { get; set; }
-    public decimal DiscountPercent { get; set; }
+    public double DiscountPercent { get; set; }
 }
