+++ cinemabe/Cinema/2-Business/Cinema.Business.DTO/Catalog/TicketTypeDTOs.cs
@@ -6,14 +6,14 @@ public class TicketTypeDTO
 {
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal BasePrice { get; set; }
+    public double BasePrice { get; set; }
     public string? Description { get; set; }
 }
 
 public class CreateTicketTypeRequest
 {
     public string Name { get; set; } = string.Empty;
-    public decimal BasePrice { get; set; }
+    public double BasePrice { get; set; }
     public string? Description { get; set; }
 }
 
@@ -21,6 +21,6 @@ public class UpdateTicketTypeRequest : IHasId
 {
     public Guid Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal BasePrice { get; set; }
+    public double BasePrice { get; set; }
     public string? Description { get; set; }
 }
