+++ cinemabe/Cinema/2-Business/Cinema.Business.DTO/Movies/ShowTimeListDTO.cs
@@ -17,5 +17,5 @@ public class ShowTimeRoomDTO
     public Guid     RoomId      { get; set; }
     public string?  RoomName    { get; set; }
     public string?  TheaterName { get; set; }
-    public decimal  BasePrice   { get; set; }
+    public double  BasePrice   { get; set; }
 }
