+++ cinemabe/Cinema/3-Data/Cinema.Data.Contracts/IInvoiceStore.cs
@@ -9,6 +9,6 @@ public interface IInvoiceStore : IGenericStore<Invoice>
     Task<Invoice?> GetByCodeAsync(string code);
     Task<(IEnumerable<Invoice> Items, int Total)> GetByUserAsync(Guid userId, int page, int pageSize);
     Task<(IEnumerable<Invoice> Items, int Total)> GetPagedAsync(InvoiceStatus? status, DateTime? from, DateTime? to, int page, int pageSize);
-    Task<decimal> GetTotalRevenueAsync(DateTime from, DateTime to);
-    Task<IReadOnlyDictionary<DateTime, decimal>> GetRevenueByDayAsync(DateTime from, DateTime to);
+    Task<double> GetTotalRevenueAsync(DateTime from, DateTime to);
+    Task<IReadOnlyDictionary<DateTime, double>> GetRevenueByDayAsync(DateTime from, DateTime to);
 }
