+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/MemberShip.cs
@@ -4,6 +4,6 @@ public class MemberShip : BaseEntity
     public string Name { get; set; } = string.Empty;
     public int MinPoints { get; set; }
     public int MaxPoints { get; set; }
-    public decimal DiscountPercent { get; set; }
+    public double DiscountPercent { get; set; }
     public ICollection<User> Users { get; set; } = new List<User>();
 }
