+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/SeatTypeTicketType.cs
@@ -3,7 +3,7 @@ public class SeatTypeTicketType
 {
     public Guid SeatTypeId { get; set; }
     public Guid TicketTypeId { get; set; }
-    public decimal PriceMultiplier { get; set; } = 1;
+    public double PriceMultiplier { get; set; } = 1;
     public SeatType SeatType { get; set; } = null!;
     public TicketType TicketType { get; set; } = null!;
 }
