+++ cinemabe/Cinema/3-Data/Cinema.Data.Entities/Entities/TicketType.cs
@@ -2,7 +2,7 @@
 public class TicketType : BaseEntity
 {
     public string Name { get; set; } = string.Empty;
-    public decimal BasePrice { get; set; }
+    public double BasePrice { get; set; }
     public string? Description { get; set; }
     public ICollection<SeatTypeTicketType> SeatTypeTicketTypes { get; set; } = new List<SeatTypeTicketType>();
     public ICollection<InvoiceTicket> InvoiceTickets { get; set; } = new List<InvoiceTicket>();
